:bug: Use response data instead of doing another request
The ConnectionDeleate used to "redo" the current request again,
by calling [webview loadRequest: [connection currentRequest]];

This can mess up stateful authentication flows.

To fix this, we wait for the responseData and load it to the webview:
[webView loadHTMLString:html baseURL:currentUrl];

The webview won't load the page content automatically when in basicAuth flow.
It does however in all other cases.

So we introduced the _in_basic_auth_flow flag, to keep track if the user is in a basic auth flow.
The _basic_authorized_failed flag could be removed, imho.