:bug: Fix auto-close handling via pattern
If a request fails with error.code == 102, the handling introduced in
e2e23df tries to open the url in an external app (in order to support
downloads and other interactions like adding calendar events).

This broke the auto close behaviour of the tabs-module, which uses the
same error.

So we exclude urls matching the pattern from the open-external-app
handling here.