RE-314: :bug: Don't send requests twice (when basicAuth is on)
If basicAuth is enabled the tabs_ConnectionDelegate is used if
the page to load is not yet in authorizationCache.

Prior this change handleRequest always return YES, which means
that the request should be performed.

But by attaching the delegate, another request is triggered.

To solve this, we stop loading by returning NO in handleRequest
and always reuse the fetched response data for later use.
(This was only be done in the basic auth challenge case before)